requirements.txt
streamlit>=1.45.1
tensorflow>=2.19.0
opencv-python-headless
pillow
numpy
openai==0.28.0
requests
